formatting

+++ esphome/components/qn8027/qn8027defs.h
@@ -49,7 +49,7 @@ static const uint8_t RDS_TEXT_MAX_SIZE = 64;
 static const uint8_t CID1_FM = 0;
 static const uint8_t CID3_QN8027 = 4;
 
-static const uint16_t XISEL_MAX = 393.75f;
+static const float XISEL_MAX = 393.75f;
 
 // TODO: enum class
 static const uint8_t FSM_STATUS_RESET = 0;
